requirements.txt
apache-airflow==2.10.0b2
apache-airflow-providers-amazon==8.27.0
apache-airflow-providers-postgres==5.11.3
pandas==2.1.4
pyTelegramBotAPI==4.22.0
environs==11.0.0
